third_party/blink/web_tests/platform/generic/accessibility/aria-hidden-children-not-in-text-from-content-expected.txt

This tests that aria-hidden elements within an element are not included in accessible name from content

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS checkbox1.name.trim() is ''
PASS checkbox2.name.trim() is 'Non-hidden descendant content'
PASS successfullyParsed is true

TEST COMPLETE
